FAB-1517 Add shake hash 256 to hash.js
The orderer selected ShakeHash 256 in the place of SHA3-512 which
would be a much slower algorithm for the same security strength
(256 bits).

so the SDK must sent a shake hash of 256 bits strength (length of
512 bits or 64 bytes) in the configuration request to the orderer
as part of constructing a channel.

This PR adds the new API to the hash primitive class hash.js, to
be used by the SDK code that creates channels.

Change-Id: I180dedc7aa23d8f6695406bc2a4164e00aed39e2
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>