SDK loads pre-provisioned users - step1
[FAB-1787] the SDK needs to support loading private keys and ecerts that
have been provisioned outside of the SDK. the crypto materials
should be saved in a "standard" format for sharing and communication
purposes.

since the peers have already adopted the "MSP configuration directory"
format, which allows for the private keys and signing certificates
to be saved, that seems to be a good choice for this.

Step 1 is to update the endorser test to perform this task. once
proved working, the same code will be used to enhance the MSP class.

Change-Id: If178109a312633dabf530cdff755f8edd41aa16b
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>