Make getUserContext() support synchronous calls
FAB-3048
Most of the time the caller code just wants to get access to the
current user context for signing stuff, yet this call is currently
asynchronous (returning a Promise) due to its support for edge
cases where the calling code either wants to switch to a different
user or is recovering from crashes.

To make the mainstream usage of this method more straightforward,
change the implementation to return synchronously by default but
can be made Promise-based if both a "name" and "checkPersistence"
are passed in.

Change-Id: Ic8b25f40aa6d80bb40b450ea0fa931c62c81d887
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>