[FAB-3569][FAB-3570] NodeSDK - work with ConfigTX
Allow users to sign the output of the configtx Tool.
Requires opening the configtx output and finding
the 'ConfigUpdate' object. This will then be an
object that the orgs will sign before submitting
to be created.

Test cases for an end to end flow
New configTX binary with new channel name so it will
not conflict with the existing mychannel.

Change-Id: I806e927b8d46971c02143b7bdace9139efceb774
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>