[FAB-3690] e2e failed due to instantiation policy
the `default` that's been used for a long time by the SDKs (node
and Java) needs to be the actual chain ID as per the LSCC API.
The recent changes with enforcing instantiation policies on the
peer side turned up this issue.

Note that until the peer side has fixed the default instantiation
policy, this will still fail because the other orgs in the channel
will still reject the proposal because each peer currently uses
their local MSP to enforce the access control.

Change-Id: I5aae54594fe4d3672315565f0cc3b61c985a91b1
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>