[FAB-5363] NodeSDK - common config part 1
The NodeSDK use of the common network configuration.
This is using the latest yaml schema as described in
the FAB. This demonstrates how the SDK would load a
configuration and build a channel instance based on the
network configuration. This starts the use of how peers
would be automatically selected.

Change-Id: I0b1479d12684aa13bc052e872fbf118166004453
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>