[FAB-8456] Document known vulnerabilties
The vulnerabilities found by the nsp (Node Security Platform) tool
are in indirect dependencies and can't currently be fixed until dependent
packages are updated; therefore, these vulnerabilities are documented
in the release notes.

In addition, the synk tool showed a vulnerability in an older version of
the restify package which is fixed by upgrading to the latest version of
restify.

Change-Id: I8e4bc204b96905cf9de246929f893230276381d9
Signed-off-by: Keith Smith <bksmith@us.ibm.com>